agent/requirements.txt
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
mcp==1.1.2
python-dotenv==1.0.0
openai>=1.55.3